fix bug with Qt >= 5.14 (fixes #10)
The problem is that for normal images Qt now resets the file device.
The handler however assumed the device will not be touched for the
handlers lifetime. This has been addressed by storing the position after
the header data and seeking the device to that position in case it
has been reset. This only happens if the device is still the same. If
the device was reset, cached frames will be deleted as well